1. 新增 harris opencv 代碼 2. focalsFromHomography 改為 double， round 改為 (int)+0.5

+++ feat_fast/getFocus/getFocus.hpp
@@ -86,7 +86,7 @@ public:
 };
 
 
-float getWarpFocal(const vector<float>& HomogMat, 
+double getWarpFocal(const vector<float>& HomogMat, 
 	size_t img1Size, size_t img2Size);
 
 void getWarpOffset(const ImgRaw &imgA, const ImgRaw &imgB,
